Refactored EnrollmentVerification login components to static page react widgets
