+++ src/applications/enrollment-verification/containers/EnrollmentVerificationIntroPage.jsx
@@ -6,7 +6,6 @@ import { useHistory } from 'react-router-dom';
 import { fetchPost911GiBillEligibility } from '../actions';
 
 import EnrollmentVerificationLoadingIndicator from '../components/EnrollmentVerificationLoadingIndicator';
-import EnrollmentVerificationLogin from '../components/EnrollmentVerificationLogIn';
 import EnrollmentVerificationPageWrapper from '../components/EnrollmentVerificationPageWrapper';
 import {
   REVIEW_ENROLLMENTS_RELATIVE_URL,
@@ -82,6 +81,13 @@ export function EnrollmentVerificationIntroPage({
     return <EnrollmentVerificationLoadingIndicator />;
   }
 
+  const EVLoginPlaceholder = (
+    <div
+      data-widget-type="view_enrollment_verification_login"
+      id="enrollment_verification_login_widget"
+    />
+  );
+
   return (
     <EnrollmentVerificationPageWrapper>
       <h1>Verify your school enrollments</h1>
@@ -90,7 +96,7 @@ export function EnrollmentVerificationIntroPage({
         continue getting paid.
       </p>
 
-      {!isLoggedIn && <EnrollmentVerificationLogin />}
+      {!isLoggedIn && EVLoginPlaceholder}
       {isLoggedIn &&
         !enrollmentVerificationFetchComplete && (
           <EnrollmentVerificationLoadingIndicator message="Loading..." />
