+++ src/applications/static-pages/widgetTypes.js
@@ -60,4 +60,6 @@ export default {
   VIEW_DEPENDENTS_CTA: 'view-dependents-CTA',
   VIEW_PAYMENT_HISTORY: 'view-payment-history',
   VIEW_TEST_AND_LAB_RESULTS_PAGE: 'view-test-and-lab-results-page',
+
+  VIEW_ENROLLMENT_VERIFICATION_LOGIN: 'view_enrollment_verification_login',
 };
